Fix frontend tests under Node 16
Node 16 starts to fail program when there is an unhandled promise. This happens in our frontend tests a lot since we are trying to fetch some API that we don't have enough permission, or when we are missing the context to provide info. This PR fixes all of the problems and make the tests to pass on Node 16.