IDOL: Remove DTI 48 (#684)
### Summary <!-- Required -->
We've hidden Games page since this PR:
https://github.com/cornell-dti/idol/pull/359.

It's been over 2 years, and game remains broken, and we have no plans to
work on it. Let's kill off the page for now to avoid keeping dead code
around. We can always find it again if we want to bring back DTI 48.

### Notion/Figma Link <!-- Optional -->

<!-- If the changes have associated Notion pages/Figma design(s), please
include the links here.-->

### Test Plan <!-- Required -->

<!-- Provide screenshots or point out the additional unit tests -->

### Notes <!-- Optional -->

<!--- List any important or subtle points, future considerations, or
other items of note. -->

### Breaking Changes <!-- Optional -->