Website: Fix Linkedin Links and Add Falsy Check for Links (#726)
### Summary <!-- Required -->
Noticed that there's a number of profiles in Alumni page that don't have
Linkedin's or have malformed URL's. I fixed the malformed URL's and add
all the missing links.

Also fixed how we handle empty links. Right now, a link of `""` would
lead to a redirect back to the current page (instead of a no-op).

### Notion/Figma Link <!-- Optional -->

https://www.notion.so/Website-Fix-Linkedin-Links-and-Fix-Check-for-Empty-Links-1450ad723ce1801a8058eeeaa05ebfd3?pvs=4


### Test Plan <!-- Required -->
Manually click through all links on alumni page, and observe that it
redirects to their LinkedIn profiles.

### Notes <!-- Optional -->

<!--- List any important or subtle points, future considerations, or
other items of note. -->

### Breaking Changes <!-- Optional -->