IDOL: Add Semester Joined and College to User Profile (#733)
### Summary <!-- Required -->
This PR adds semester joined and college to the user profile. Semester
joined is used to keep track of who is a newbie, for coffee chat
requirement purposes. College is used on the new website on the pie
chart. Semester joined is visible but not editable on the user side --
it should be configured by admin. College is editable on the user side.

### Notion/Figma Link <!-- Optional -->

[Notion](https://www.notion.so/Website-IDOL-Add-Semester-Joined-and-College-Fields-to-IdolMember-1440ad723ce18056b66ec5436eb56959)

### Test Plan <!-- Required -->
Editing the fields, verifying values persisted in database.


https://github.com/user-attachments/assets/70723fef-cb8e-48c4-9ba8-e6373739b185

### Notes <!-- Optional -->
* Semester joined and college have not been populated in the database.
* Though I have marked both fields as required, they can be undefined
for now for backward compatibility.

---------

Co-authored-by: Andrew Chen <andrew032012@gmail.com>