created function to fix links in candidate decider (#728)
### Summary <!-- Required -->

I created a function that checks links in Candidate Decider and edits
them so that they are proper links. For example, it adds to https:// to
a link such as www.linkedin.com/in/andrew-x-chen/ and adds http://www.
to linkedin.com/in/andrew-x-chen/. It will also transfrom a linked in
username into a link, for example, andrew-x-chen to
http://www.linkedin.com/in/andrew-x-chen/. It does the same for github
links. It also parses links out of badly formatted text such as "This is
not my portfolio but it is close!
https://jason-mun-fallback-website.vercel.app/."

### Notion/Figma Link <!-- Optional -->

[https://www.notion.so/Candidate-Decider-Make-Link-Reading-More-Robust-1430ad723ce18005b836d0e6adf5732c](url)

### Test Plan <!-- Required -->

Here is a screen recording of the testing that I did of various links.
There was a strange error at the end of the recording but I don't think
its related to the links and it went away on the second try...


https://github.com/user-attachments/assets/dbd75f37-25e7-4def-8e78-8b7d9107908e

---------

Co-authored-by: Andrew Chen <59291082+andrew032011@users.noreply.github.com>