Pull from IDOL Images: Compress Images to < 300 KB (#739)
### Summary <!-- Required -->
We previously didn't have any image compression logic, which required us
to manually compress images when they exceed the limit.

Make the Pull from IDOL Images script compress images to < 300 KB. We
can also revisit this if it makes the images too low quality/blurry.

We may consider overwriting the bucket's images with the compressed
version, but only if we verify that the quality is not significantly
diminished. For now, we should keep the images in the bucket as a source
of truth for what the original image looked like.

### Notion/Figma Link <!-- Optional -->

https://www.notion.so/Website-Compress-Images-in-Pull-from-IDOL-Images-Job-1410ad723ce18056a2f1f224afae14d1?pvs=4

### Test Plan <!-- Required -->
Pull from IDOL Images PR: https://github.com/cornell-dti/idol/pull/729


### Notes <!-- Optional -->

<!--- List any important or subtle points, future considerations, or
other items of note. -->

### Breaking Changes <!-- Optional -->