New Website: Ensure Proper Use of Headers (#755)
### Summary <!-- Required -->
Ensure:
- Each page has a `<h1>` tag (always the hero section only)
- Don't skip over headers (i.e. every page now has `<h1>` for header,
`<h2>` for subheader, and `<h3>` for headers below, and if there's a
header below that, there's a `<h4>`. There aren't any usages of `<h5>`'s
or `<h6>`'s anymore.

There should be no difference in the UI from this change.

### Notion/Figma Link <!-- Optional -->
Resolves problem 2 of https://github.com/cornell-dti/idol/issues/734.


https://www.notion.so/Address-Code-Audit-Comments-14b0ad723ce1802fa2f0c633511d8de3?pvs=4


### Test Plan <!-- Required -->
Verify no UI changes are introduced by this change.

Grep for `<h1>`, `<h2>`, `<h3>`, `<h4>`, `<h5>`, `<h6>` and ensure that
it meets the standards mentioned above.

### Notes <!-- Optional -->

<!--- List any important or subtle points, future considerations, or
other items of note. -->

### Breaking Changes <!-- Optional -->