Website Redesign: Slideshow fix (#770)
### Summary <!-- Required -->

This PR fixes the Slideshow delay bug on the Home page. Instead of
relying on the dynamic calculation for the width from using the
useScreenSize() hook, I'm simply hiding the element using the tailwind
css utilities so we don't have to calculate it on the mount. Flex-col
wasn't needed since I didn't realize that the order of the columns
wasn't the same haha!

- [x] Fixed weird delay issue so it should be good! 
- [x] Changed the Red blob as well so it didn't rely on width and so is
also present on mount!
- [x] Adjusted Icon widths and utilized scale for styling instead of
width

### Notion/Figma Link <!-- Optional -->

[Slideshow Delay
bug](https://www.notion.so/New-Website-Slideshow-has-a-delay-when-loading-1400ad723ce1801d99a4e02db3b9d760?pvs=4)

### Test Plan <!-- Required -->


https://github.com/user-attachments/assets/cbbd1cfb-f6df-4f81-80d7-a0cbdf2fe3fa