New Website: Set up Google Analytics (#772)
### Summary <!-- Required -->
Largely copy paste from https://github.com/cornell-dti/idol/pull/283,
adapted to work with `app/` directory instead of the older `pages/`
directory model.

Configured a new stream on our google analytics account for new website
and set the URL to be new.cornelldti.org while we're still pre-launch.

### Notion/Figma Link <!-- Optional -->

https://www.notion.so/Set-up-google-analytics-for-new-website-d8d425162cd24b29b48a248d948a8665?pvs=4

### Test Plan <!-- Required -->
We can test this by hitting the website a bunch of times across a few
days to see if any data is received. It can take up to 48 hours for data
to begin showing.

### Notes <!-- Optional -->

<!--- List any important or subtle points, future considerations, or
other items of note. -->

### Breaking Changes <!-- Optional -->