Coffee Chat: Admin Can View Individual Member Coffee Chat Bingo Boards (#774)
### Summary <!-- Required -->

I added a dropdown in the admin view that allows the admin to see the
bingo board for each member. When the admin selects a specific member,
the bingo board that shows all of the current pending coffee chats
becomes the specific member's bingo board. I also created a
CoffeeChatsBingoBoard component that is now used by CoffeeChatsDashboard
and CoffeeChats.

This pull request still requires more revision as there are some
problems I have run into and not been able to solve. So if a reviewer
could provide me with some advice on how to fix the problems that I will
detail below I would greatly appreciate it.


### Notion/Figma Link <!-- Optional -->


https://www.notion.so/Coffee-Chat-Admin-Can-View-Individual-Member-Coffee-Chat-Bingo-Boards-1440ad723ce18003ad6cc3a266b311ab

### Test Plan <!-- Required -->

Based on my tests, there are some issue occurring as I mention below in
the notes. But this is how it is currently functioning.


https://github.com/user-attachments/assets/53fa7f01-2aff-4544-a9a0-54d526542d83


### Notes <!-- Optional -->

Currently, when a specific member is selected, the bingo board becomes
smaller. I haven't been able to figure out how to ensure that the bingo
boards stay the same size.
Also, another issue is that when other members are selected, it still
only shows my - Juju Crane's - bingo board. It appears that the member
that the get request is being requested on is the correct member as it
shows the correct email in the request however, for some reason the
response is returning my coffee chat data?

<img width="1440" alt="Screenshot 2024-12-06 at 1 04 26 PM"
src="https://github.com/user-attachments/assets/6ebfd77c-c42b-4885-83b3-0dc2855784d6">

---------

Co-authored-by: Andrew Chen <andrew032012@gmail.com>