remove deprecated workspaces (#987)
### Summary <!-- Required -->
The builds were failing in #986 because we were using a deprecated
version of `node-gyp` (v5.1.1), found in the dependency path of
`compress-images` in `image-cli`. `image-cli` was used to resize images
for the website, but now that is done in the workflow script (see
`/backend/scripts/pull-from-idol-images.ts`).

Along with this workspace, the old versions of the DTI website are
deprecated and no longer maintained. As such, they should all be removed
to prevent legacy dependency problems, reduce build times, and decrease
technical debt.

### Test Plan <!-- Required -->
```
yarn
```
Verify that the CD build works.

### Notes <!-- Optional -->
Error log
```
9:17:39 PM: gyp ERR! UNCAUGHT EXCEPTION
9:17:39 PM: gyp ERR! stack TypeError: Cannot assign to read only property 'cflags' of object '#<Object>'
9:17:39 PM: gyp ERR! stack     at createConfigFile (/opt/build/repo/node_modules/node-gyp/lib/configure.js:118:21)
9:17:39 PM: gyp ERR! stack     at /opt/build/repo/node_modules/node-gyp/lib/configure.js:85:9
9:17:39 PM: gyp ERR! stack     at /opt/build/repo/node_modules/mkdirp/index.js:30:20
9:17:39 PM: gyp ERR! stack     at FSReqCallback.oncomplete (node:fs:188:23)
9:17:39 PM: gyp ERR! System Linux 5.10.233-224.894.amzn2.x86_64
9:17:39 PM: gyp ERR! command "/opt/buildhome/.nvm/versions/node/v20.19.3/bin/node" "/opt/build/repo/node_modules/.bin/node-gyp" "rebuild"
9:17:39 PM: gyp ERR! cwd /opt/build/repo/image-cli/node_modules/sharp
9:17:39 PM: gyp ERR! node -v v20.19.3
9:17:39 PM: gyp ERR! node-gyp -v v5.1.1
9:17:39 PM: gyp ERR! Node-gyp failed to build your package.
9:17:39 PM: gyp ERR! Try to update npm and/or node-gyp and if it does not help file an issue with the package author.
```