Interview Scheduler Bug Bashing Fixes (#932)
### Summary <!-- Required -->
Bug Fixes:
* Reinforced permission checking for update interview slot endpoint.
Members and applicants should not be able to override each other, and
only one user may get an approval message when signing up for the same
slot on the frontend.
* Leads are able to bypass this checking. This means that although IDOL
subteam members are able to edit the time slot to whomever they want,
the checks will still be done. Direct database edits will need to be
done to make these changes if not a lead.
* Fixed an issue where the backend would request the redacted scheduler
information.
* Fixed an issue where applicants cannot sign up because the returned
scheduler from the endpoint redacts all applicants.
*  Reformatted side panel when editing.
*  Include date on invite card.
*  Allow admins to schedule interviews on non-consecutive days.

### Test Plan <!-- Required -->
Testing during worksesh!

### Notes <!-- Optional -->
Applicants should have much more robust permission checking, especially
since this is open-source. This will be one of the only endpoints that
allow non-IDOL members to interact with IDOL. It is imperative to make
sure that outside users may not make unauthorized endpoint requests.