Combine 'Edit Profile' and 'Edit Profile Image' (#1086)
### Summary <!-- Required -->

<!-- Provide a general summary of your changes in the Title above -->
<!-- Itemize bug fixes, new features, and other changes -->
<!-- Feel free to break this into sub-sections, i.e. features, fixes,
etc. -->
<!-- Some examples are shown below. -->

This pull request is to combine the two pages - 'Edit Profile' and 'Edit
Profile Image' - into just one. Essentially, the bulk of this is
migrating the code in 'Edit Profile Image' into 'Edit Profile' and
making sure it is integrated into the page properly.

- [x] migrated 'Edit Profile Image' Code
- [x] update tests for 'Edit Profile'
- [x] fix the layout for the profile image card

### Linear Link <!-- Optional -->

[Combine Edit Profile and Edit Profile
Image](https://linear.app/cornelldti/issue/IDOL-39/combine-edit-profile-and-edit-profile-image
)
<!-- If the changes have associated Notion pages/Figma design(s), please
include the links here.-->

### Test Plan <!-- Required -->

<!-- Provide screenshots or point out the additional unit tests -->

1. Go to homepage of Idol and ensure that 'Edit Profile Image' is not
there.
2. Go to 'Edit Profile' and ensure that your profile image is there,
along with a button to edit it.
3. Ensure you can open the editor both by clicking the button and your
profile image.
4. Update your image (play around with the modifications) and make sure
it saves

### Notes <!-- Optional -->

Need to update 'Edit Profile' tests to test for this addition
<!--- List any important or subtle points, future considerations, or
other items of note. -->