Updated coffee chat suggestions script (#1068)
### Summary <!-- Required -->

- Updated semester from Spring 2025 to Fall 2025
- Removed dev advisor checks since that is no longer a coffee chat
category
- Ran script to update new coffee chat suggestions
- Adjusted tests to reflect these changes

### Linear
[coffee chat
suggestions](https://linear.app/cornelldti/issue/IDOL-38/coffee-chat-suggestions)

### Test plan
- Go to coffee chat bingo board
- Check that each category has populated suggestions, no duplicate names
(newbies should be easy to verify for accuracy of suggestions)

### Note 
I wanted to make a pr to merge the script since we would need to use it
every semester (originally from Chris's branch).

Also right now, we need to manually check for members who make multiple
submissions to the google form or submit their name in lowercase
(generally just validate the data). Do we want to handle those cases in
the script? Or since it's not too much work it's fine?